@@ -1,28 +1,19 @@
 <template>
-  <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png">
-    <HelloWorld msg="Welcome to Your Vue.js App"/>
-  </div>
-</template>
+  <Header />
+  <router-view/>
+  <Footer />
 
-<script>
-import HelloWorld from './components/HelloWorld.vue'
+ </template>
 
-export default {
-  name: 'App',
-  components: {
-    HelloWorld
-  }
-}
-</script>
+ <script>
+ import Header from './components/Header.vue'
+ import Footer from './components/Footer.vue'
 
-<style lang="scss">
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
-}
-</style>
+ export default {
+   name: 'App',
+   components: {
+     Header,
+     Footer,
+   }
+ }
+ </script>
